requirements.txt
pypdf2
langchain-openai
langchain-community
python-dotenv
tiktoken
streamlit
pymupdf